fix: done

+++ src/lib/gitMethod.ts
@@ -1,6 +1,6 @@
-import { cmd } from '../cmd'
+import { cmd } from './cmd'
 import { ExtConfig, workspaceWithConfig } from '../main'
-import { chcheDirName, repoName } from '../const'
+import { chcheDirName, repoName } from '../constant'
 import { existsSync, mkdirSync } from 'fs'
 import dayjs = require('dayjs')
 
@@ -23,23 +23,33 @@ export class GitMethods {
   }
 
   async initRepo ({ config: { remote, branch } }: workspaceWithConfig) {
-    if (!existsSync(`${this.execPath}/${chcheDirName}`)) {
-      mkdirSync(`${this.execPath}/${chcheDirName}`)
-    }
     // @see: https://askubuntu.com/a/483192；https://unix.stackexchange.com/questions/6393/how-do-you-move-all-files-including-hidden-from-one-directory-to-another
     const match = '{*,.[^auto]*}'
-    await this.cmd(`mv ./${match} ${chcheDirName}`).catch(() => {
-      // do nothing
-    })
+
     await this.cmd('git init')
     await this.addRemote(remote)
-    await this.cmd(`git fetch ${repoName}`)
-    await this.cmd(`git checkout -f ${repoName}/${branch} -b master`)
 
-    await this.cmd(`cp -R ${chcheDirName}/${match} ./`).catch(() => {
-      // do nothing
-    })
-    await this.cmd(`rm -rf ${chcheDirName}`)
+    await this.cmd(`git fetch ${repoName}`)
+    // 先行判断是否存在目标分支
+    debugger
+    const branchName = await this.cmd(`git branch -la | grep ${repoName}/${branch}`)
+    debugger
+    if (branchName) {
+      if (!existsSync(`${this.execPath}/${chcheDirName}`)) {
+        mkdirSync(`${this.execPath}/${chcheDirName}`)
+      }
+      await this.cmd(`mv ./${match} ${chcheDirName}`).catch(() => {
+        // do nothing
+      })
+      await this.cmd(`git checkout -f ${repoName}/${branch} -b ${branch}`).catch(_ => {
+        // 是一个空项目，或者目标分支不存在
+        // do nothing
+      })
+      await this.cmd(`cp -R ${chcheDirName}/${match} ./`).catch(() => {
+        // do nothing
+      })
+      await this.cmd(`rm -rf ${chcheDirName}`)
+    }
   }
 
   async addRemote (remote: string) {
@@ -50,20 +60,16 @@ export class GitMethods {
     await this.cmd(`git remote set-url ${repoName} ${remote}`)
   }
 
-  async syncRemote (branch: string) {
+  async syncRemote (branch: string): Promise<boolean> {
     return await this.cmd(`git pull ${repoName} ${branch}`)
-      .then(() => {
-        return true
-      })
-      .catch(() => {
-        return false
-      })
+      .then(_ => true)
+      .catch(_ => false)
   }
 
   async commitAndPush ({ branch }: ExtConfig) {
     await this.cmd('git add .')
     await this.cmd(`git commit -m commit_by_vsc-auto-diary-ext_${dayjs().format('YYYY/MM/DD_HH:mm')}`)
-    return await this.cmd(`git push auto-diary HEAD:${branch}`)
+    return await this.cmd(`git push auto-diary ${branch}:${branch}`)
       .then(() => {
         return Promise.resolve(true)
       })
